Cleanup dead code (#5)
* Remove unused _on_Options method

The Game._on_Options method is never called because it's missing the
required boolean argumend.  If the argument is added, the method is
called but it assigns the non-existent property $OptionsMenu.show,
causing an error.  Probably old code still sitting around.

* Delete duplicate files

These files are not used, probably sitting around after a change in
naming convention.